restructure and tags, users, genres saving to db
